I think Sterling missed a commit, compiles on Vogon now
git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@507 342171fe-4ba3-422f-b357-09b6f24955fe

+++ Items/AsteroidShip.cpp
@@ -932,7 +932,7 @@ void AsteroidShip::drawCrosshair() {
       glEnd();
       */
    if (currentView == VIEW_FIRSTPERSON_SHIP) {
-      glTranslatef(getAimX(), getAimY(), 0.0);
+      glTranslatef((GLfloat)getAimX(),(GLfloat) getAimY(),(GLfloat)0.0f);
    }
    static GLUquadricObj *outer;
    static GLUquadricObj *inner;
